+++ RayTracerChallenge/Chap.6/Material.swift
@@ -10,10 +10,10 @@ import Foundation
 
 struct Material: Equatable {
 	var color = Color.white
-	var ambient: Float = 0.1
-	var diffuse: Float = 0.9
-	var specular: Float = 0.9
-	var shininess: Float = 200.0
+	var ambient: Double = 0.1
+	var diffuse: Double = 0.9
+	var specular: Double = 0.9
+	var shininess: Double = 200.0
 	
 	static func == (lhs: Material, rhs: Material) -> Bool {
 		guard
